pyproject.toml
[project]
name = "auto-check"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jaeyoon492", email = "audwodbs492@gmailcom"},
]
dependencies = ["requests>=2.32.3", "selenium>=4.29.0", "python-dotenv>=1.0.1", "discord>=2.3.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "python main.py"
